requirements.txt
Babel==2.6.0
certifi==2019.3.9
chardet==3.0.4
DateTime==4.3
defusedxml==0.6.0
Django==2.2
django-allauth==0.39.1
django-cors-headers==2.4.0
django-filter==2.1.0
django-phone-field==1.7.0
django-phonenumber-field==2.3.1
django-rest-auth==0.9.3
django-stubs==0.12.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.4.2
idna==2.8
mypy==0.701
mypy-extensions==0.4.1
oauthlib==3.0.1
phonenumbers==8.10.10
Pillow==6.0.0
psycopg2==2.8.2
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.4
typing-extensions==3.7.2
urllib3==1.24.2
whitenoise==3.3.1
zope.interface==4.6.0
